requirements.txt
pyopengl
pygame
numpy
